requirements.txt
anyio==3.3.0
appdirs==1.4.4
certifi==2021.5.30
charset-normalizer==2.0.4
dacite==1.6.0
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
onlinesimru==1.0.17
pyee==8.2.2
pyppeteer==0.2.6
python-dateutil==2.8.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tqdm==4.62.2
urllib3==1.26.6
uvloop==0.14.0
websockets==9.1
zipp==3.5.0
